Project complete, all conditions seem to be met
Have tested on both small and large files, file not existing and multiple files consecutively. Seems to function as imagined. Could be sped up by removing the various print statements used for debugging, as even setting DEBUG to false, the if statements still slow down the program and Junit tests would be a more functional way of testing to see if the client works correctly anyways.